- Added back the code to allow some variation to the players' shades when   players are on teams. - Set TEAM_None back to 255. Since a player's team has already been accessible   through ACS, needlessly redefining this is a bad thing to do, since it can   break existing maps. 255 different teams should still be more than enough. - Fixed: At certain resolutions, there was a one pixel row between the status   bar and the rest of the screen, thanks to rounding error. - Added automatic batching of quads to D3DFB. Screens with a lot of text are   ever-so-slightly faster now, though still only about half the speed of   sofware-only text. I suppose the only way to see a marked improvement is   going to be by stuffing multiple glyphs in a single texture. - Fixed: Crosshairgrow's animation was not framerate-independent.
SVN r668 (trunk)

+++ src/b_game.cpp
@@ -588,7 +588,7 @@ bool DCajunMaster::LoadBots ()
 			case BOTCFG_TEAM:
 				{
 					char teamstr[16];
-					unsigned int teamnum;
+					BYTE teamnum;
 
 					SC_MustGetString ();
 					if (IsNum (sc_String))
